+++ app/(main)/location/components/MapSection/index.tsx
@@ -8,7 +8,6 @@ import VehicleMarker from '@/app/(main)/location/components/VehicleMarker'
 import Map from '@/components/domain/map/Map'
 import { useMapStatus } from '@/hooks/useCurrentMapStatus'
 import { clusterService } from '@/lib/apis/cluster'
-// import ClusteringData from '@/mock/clustering.json'
 import { LatLng } from '@/types/location'
 import { ClusterPoint, MapState } from '@/types/map'
 import { VehicleInfoModel } from '@/types/vehicle'
