+++ hooks/useSearchVehicle.ts
@@ -3,10 +3,10 @@ import { useState } from 'react'
 import { useModal } from '@/hooks/useModal'
 import { vehicleService } from '@/lib/apis'
 import { validateVehicleNumber } from '@/lib/utils/validation'
-import { VehicleDateModel } from '@/types/vehicle'
+import { VehicleOperationPeriodModel } from '@/types/vehicle'
 
 export const useSearchVehicle = (vehicleNumber: string = '') => {
-    const [searchedVehicle, setSearchedVehicle] = useState<VehicleDateModel>()
+    const [searchedVehicle, setSearchedVehicle] = useState<VehicleOperationPeriodModel>()
     const [searchableDates, setSearchableDates] = useState({ firstDateAt: '', lastDateAt: '' })
 
     const { isOpen, modalMessage, closeModal, showMessage } = useModal()
@@ -20,36 +20,24 @@ export const useSearchVehicle = (vehicleNumber: string = '') => {
         }
 
         try {
-            // const response = await vehicleAPI.fetchVehicleData(vehicleNumber)
-            const response = {
-                vehicleId: '1',
-                vehicleNumber: '123가1234',
-                searchableDates: {
-                    firstDateAt: '2024-02-02T11:11:11',
-                    lastDateAt: '2024-02-02T11:11:11',
-                },
-            }
+            const response = await vehicleService.getVehicleOperationPeriod(vehicleNumber)
 
-            await vehicleService.getVehicleOperationPeriod(vehicleNumber)
-
-            // TODO: 등록되지 않는 차량 API 에러 메세지로 대체
-            if (validation.value !== response.vehicleNumber) {
+            if (!response.isValid) {
                 showMessage('등록되지 않은 차량번호입니다.')
                 return
             }
 
-            const vehicleData = {
-                vehicleId: response.vehicleId,
-                vehicleNumber: response.vehicleNumber,
-                // searchableDates: {
-                //     firstDateAt: response.firstDateAt,
-                //     lastDateAt: response.lastDateAt,
-                // },
-                searchableDates: response.searchableDates,
+            const vehicleOperationPeriodInfo = {
+                vehicleId: response.value.vehicleId,
+                vehicleNumber: response.value.vehicleNumber,
+                searchableDates: {
+                    firstDateAt: response.value.firstDataAt,
+                    lastDateAt: response.value.lastDataAt,
+                },
             }
 
-            setSearchedVehicle(vehicleData)
-            setSearchableDates(vehicleData.searchableDates)
+            setSearchedVehicle(vehicleOperationPeriodInfo)
+            setSearchableDates(vehicleOperationPeriodInfo.searchableDates)
         } catch (error) {
             console.error(error)
             showMessage('차량 정보를 불러오는데 실패했습니다')
